added creation of thumbnail for a selected Video from picker. This for createPost
